loader: extract Lua chunk compilation into a chunks module
Second piece of breaking up #731 into independently reviewable steps:
loader's loadstr helper (compiling a Lua string with the Wowless
stack-taint branch) moves wholesale into a new chunks module as
LoadChunk. loader.lua takes chunks as a dependency and calls
chunks.LoadChunk at all three former call sites; the compile semantics
are unchanged.
